% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyph.support-ru.R
\name{hyph.support.ru}
\alias{hyph.support.ru}
\title{Language support for Russian}
\usage{
hyph.support.ru()
}
\description{
This function adds support for Russian to the sylly package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular,
      a new set of hyphenation patterns is being added (see \code{\link{hyph.ru}}).
To use the patterns with \code{\link[sylly:hyphen]{hyphen}}, simply use the abbreviation:
\itemize{
  \item {\code{"ru"}} {--- Russian hyphenation patterns}
}
}
\seealso{
\code{\link[sylly:hyphen]{hyphen}},
   \code{\link[sylly:set.hyph.support]{set.hyph.support}}
}
