% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sylly.ru-package.R
\docType{package}
\name{sylly.ru-package}
\alias{sylly.ru-package}
\title{The sylly.ru Package}
\description{
Language Support for 'sylly' Package: Russian.
}
\details{
\tabular{ll}{
Package: \tab sylly.ru\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2017-06-10\cr
Depends: \tab R (>= 3.1),sylly\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://reaktanz.de/?c=hacking&s=koRpus\cr
}

Adds support for the Russian language to the 'sylly' package. Due to some restrictions on
CRAN, the full package sources are only available from the project homepage. To ask for
help, report bugs, suggest feature improvements, or discuss the global development of the
package, please consider subscribing to the koRpus-dev mailing list
(<http://korpusml.reaktanz.de>).
}
\author{
Meik Michalke
}
\keyword{package}
